Small style updates.
The date and author should be in the Din font. Can we make the grey #6d6e71.
The “you might also like” plug in is missing at the bottom of each entry.
The rest of the pages on the site have different formatting.
Text under “About Jason”: please make that text #3a3a3c and 15px with spacing 20px.
Make the Join Jason in his upcoming events text in the Din font.
On the right column, the double red lines can we make a tad closer together.
On the mobile version there is a large white space below the logo.